+++ test-utils/src/latestDependencies/java/com/revenuecat/purchases/utils/stubsBillingClient4.kt
@@ -1,6 +1,7 @@
 package com.revenuecat.purchases.utils
 
 import com.android.billingclient.api.Purchase
+import org.json.JSONArray
 
 fun stubGooglePurchase(
     productIds: List<String> = listOf("com.revenuecat.lifetime"),
@@ -16,7 +17,7 @@ fun stubGooglePurchase(
     {
         "orderId": "$orderId",
         "packageName":"com.revenuecat.purchases_sample",
-        "productId":${productIds[0]},
+        "productIds":${JSONArray(productIds)},
         "purchaseTime":$purchaseTime,
         "purchaseState":${if (purchaseState == 2) 4 else 1},
         "purchaseToken":"$purchaseToken",
@@ -24,3 +25,17 @@ fun stubGooglePurchase(
     }
         """.trimIndent(), signature
 )
+
+fun stubPurchaseHistoryRecord(
+    productIds: List<String> = listOf("monthly_intro_pricing_one_week"),
+    purchaseTime: Long = System.currentTimeMillis(),
+    purchaseToken: String = "abcdefghijkcopgbomfinlko.AO-J1OxJixLsieYN08n9hV4qBsvqvQo6wXesyAClWs-t7KnYLCm3-" +
+        "q6z8adcZnenbzqMHuMIqZ9kQ4KebT_Bge6KfZUhBt-0N0U0s71AEwFpzT7hrtErzdg",
+    signature: String = "signature${System.currentTimeMillis()}"
+): PurchaseHistoryRecord = PurchaseHistoryRecord("""
+            {
+                "productIds": ${JSONArray(productIds)},
+                "purchaseTime": $purchaseTime,
+                "purchaseToken": "$purchaseToken"
+            }
+        """.trimIndent(), signature)
